move browse tests to support new project structure
